Extract file operations to module
